+++ tests/test_parser.py
@@ -120,6 +120,39 @@ from dotenv.parser import Binding, Original, parse_stream
             ],
         ),
         (
+            "a= #b",
+            [
+                Binding(
+                    key="a",
+                    value="",
+                    original=Original(string="a= #b", line=1),
+                    error=False,
+                )
+            ],
+        ),
+        (
+            "a=  # b",
+            [
+                Binding(
+                    key="a",
+                    value="",
+                    original=Original(string="a=  # b", line=1),
+                    error=False,
+                )
+            ],
+        ),
+        (
+            "a=#b",
+            [
+                Binding(
+                    key="a",
+                    value="#b",
+                    original=Original(string="a=#b", line=1),
+                    error=False,
+                )
+            ],
+        ),
+        (
             "a=b c",
             [
                 Binding(